requirements.txt
typer
inquirer
yaspin
pyfiglet
tabulate
rich
rich-gradient
python-dotenv
openai   
anthropic   
google-genai
xai-sdk
